Add IsEmail::Validators::Parser and its machinery
This change ports the parsing behavior from pyIsEmail over to Ruby. The
behavior was originally defined by Dominic Sayers and ported to Python
by me. This port is almost a line-by-line port, but with some changes to
make it slightly less verbose by using constructs that Python doesn't
have.

It also includes a comprehensive test suite that is metaprogrammed from
an included XML database. Each example includes information and credits
for the author of the test.